requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.9.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
configparser==7.2.0
cryptography==44.0.2
cx_Oracle==8.3.0
et-xmlfile==1.1.0
Flask==3.1.0
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.0.5
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybitrix24==1.1.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.1.3
yarl==1.9.4
sshtunnel==0.4.0
python-ldap==3.4.0
oracledb